+++ src/builtin_parser/runner/value.rs
@@ -1,25 +1,22 @@
+//! Runtime values
+
 use std::collections::HashMap;
-use std::fmt::Debug;
+
+use bevy::reflect::{DynamicStruct, DynamicTuple, PartialReflect};
+use logos::Span;
 
 use crate::builtin_parser::number::Number;
-use crate::builtin_parser::{
-    BRIGHT_YELLOW, Environment, GREEN, NumberKind, RED, StrongRef, UniqueRc, YELLOW,
-};
+use crate::builtin_parser::{Spanned, UniqueRc};
 
-use super::super::Spanned;
-use super::environment::FunctionParam;
 use super::error::EvalError;
-use super::reflection::{CreateRegistration, IntoResource};
+use super::reflection::IntoResource;
 use super::unique_rc::WeakRef;
 
-use bevy::ecs::world::World;
-use bevy::reflect::{
-    DynamicStruct, DynamicTuple, GetPath, PartialReflect, ReflectRef, TypeInfo, TypeRegistration,
-    VariantInfo, VariantType,
-};
+pub mod format;
+pub mod function_param;
+pub mod kind;
 
-use kinded::{Kind, Kinded};
-use logos::Span;
+pub use kind::ValueKind;
 
 /// A runtime value
 #[derive(Debug, Clone)]
@@ -100,411 +97,10 @@ impl Value {
             Value::Resource(_) => Err(EvalError::CannotReflectResource(span)),
         }
     }
-
-    /// Attempts to format this [`Value`].
-    ///
-    /// Returns an error if the [`Value`] is a reference to moved data.
-    pub fn try_format(
-        &self,
-        span: Span,
-        world: &World,
-        registrations: &[&TypeRegistration],
-    ) -> Result<String, EvalError> {
-        const TAB: &str = "    ";
-        match self {
-            Value::None => Ok(format!("()")),
-            Value::Number(number) => Ok(format!("{number}")),
-            Value::Boolean(bool) => Ok(format!("{YELLOW}{bool}{YELLOW:#}")),
-            Value::String(string) => Ok(format!("{GREEN}\"{string}\"{GREEN:#}")),
-            Value::Reference(reference) => match reference.upgrade() {
-                Some(rc) => Ok(rc.borrow().try_format(span, world, registrations)?),
-                _ => Err(EvalError::ReferenceToMovedData(span)),
-            },
-            Value::Object(map) => {
-                let mut string = String::new();
-                string.push('{');
-                for (key, value) in map {
-                    string += &format!(
-                        "\n{TAB}{key}: {},",
-                        value.borrow_inner().borrow().try_format(
-                            span.clone(),
-                            world,
-                            registrations
-                        )?
-                    );
-                }
-                if !map.is_empty() {
-                    string.push('\n');
-                }
-                string.push('}');
-                Ok(string)
-            }
-            Value::StructObject { name, map } => {
-                let mut string = String::new();
-                string += &format!("{name} {{");
-                for (key, value) in map {
-                    string += &format!(
-                        "\n{TAB}{key}: {},",
-                        value.borrow_inner().borrow().try_format(
-                            span.clone(),
-                            world,
-                            registrations
-                        )?
-                    );
-                }
-                if !map.is_empty() {
-                    string.push('\n');
-                }
-                string.push('}');
-                Ok(string)
-            }
-            Value::Tuple(tuple) => {
-                let mut string = String::new();
-                string.push('(');
-                for element in tuple.iter() {
-                    string += &format!(
-                        "\n{TAB}{},",
-                        element.value.borrow_inner().borrow().try_format(
-                            span.clone(),
-                            world,
-                            registrations
-                        )?
-                    );
-                }
-                if !tuple.is_empty() {
-                    string.push('\n');
-                }
-                string.push(')');
-                Ok(string)
-            }
-            Value::StructTuple { name, tuple } => {
-                let mut string = String::new();
-                string.push_str(name);
-                string.push('(');
-                for element in tuple.iter() {
-                    string += &format!(
-                        "\n{TAB}{},",
-                        element.value.borrow_inner().borrow().try_format(
-                            span.clone(),
-                            world,
-                            registrations
-                        )?
-                    );
-                }
-                if !tuple.is_empty() {
-                    string.push('\n');
-                }
-                string.push(')');
-                Ok(string)
-            }
-            Value::Resource(resource) => Ok(fancy_debug_print(resource, world, registrations)),
-        }
-    }
-    // /// Returns the kind of [`Value`] as a [string slice](str).
-    // /// You may want to use [`natural_kind`](Self::natural_kind)
-    // /// instead for more natural sounding error messages
-    // pub const fn kind(&self) -> &'static str {
-    //     match self {
-    //         Value::None => "none",
-    //         Value::Number(number) => number.as_str(),
-    //         Value::Boolean(..) => "boolean",
-    //         Value::String(..) => "string",
-    //         Value::Reference(..) => "reference",
-    //         Value::Object(..) => "object",
-    //         Value::StructObject { .. } => "struct object",
-    //         Value::Tuple(..) => "tuple",
-    //         Value::StructTuple { .. } => "struct tuple",
-    //         Value::Resource(..) => "resource",
-    //     }
-    // }
-}
-
-#[derive(Clone, Copy, Debug, PartialEq, Eq)]
-pub enum ValueKind {
-    None,
-    Number(NumberKind),
-    AnyNumber,
-    Boolean,
-    String,
-    Reference,
-    Object,
-    StructObject,
-    Tuple,
-    StructTuple,
-    Resource,
-}
-impl Kinded for Value {
-    type Kind = ValueKind;
-
-    fn kind(&self) -> Self::Kind {
-        match self {
-            Self::None => ValueKind::None,
-            Self::Number(number) => ValueKind::Number(number.kind()),
-            Self::Boolean(..) => ValueKind::Boolean,
-            Self::String(..) => ValueKind::String,
-            Self::Reference(..) => ValueKind::Reference,
-            Self::Object(..) => ValueKind::Object,
-            Self::StructObject { .. } => ValueKind::StructObject,
-            Self::Tuple(..) => ValueKind::Tuple,
-            Self::StructTuple { .. } => ValueKind::StructTuple,
-            Self::Resource(..) => ValueKind::Resource,
-        }
-    }
-}
-impl Kind for ValueKind {
-    fn all() -> &'static [Self] {
-        unimplemented!()
-    }
-}
-impl ValueKind {
-    #[must_use]
-    pub const fn as_str(&self) -> &'static str {
-        match self {
-            Self::None => "none",
-            Self::Number(number) => number.as_str(),
-            Self::AnyNumber => "float/integer",
-            Self::Boolean => "boolean",
-            Self::String => "string",
-            Self::Reference => "reference",
-            Self::Object => "object",
-            Self::StructObject => "struct object",
-            Self::Tuple => "tuple",
-            Self::StructTuple => "struct tuple",
-            Self::Resource => "resource",
-        }
-    }
-    /// Returns the kind of [`Value`] as a [string slice](str) with an `a` or `an`  prepended to it.
-    /// Used for more natural sounding error messages.
-    pub const fn as_natural(&self) -> &'static str {
-        match self {
-            ValueKind::None => "nothing",
-            ValueKind::Number(number) => number.as_natural(),
-            ValueKind::AnyNumber => "any number",
-            ValueKind::Boolean => "a boolean",
-            ValueKind::String => "a string",
-            ValueKind::Reference => "a reference",
-            ValueKind::Object => "a object",
-            ValueKind::StructObject => "a struct object",
-            ValueKind::Tuple => "a tuple",
-            ValueKind::StructTuple => "a struct tuple",
-            ValueKind::Resource => "a resource",
-        }
-    }
-}
-
-impl std::fmt::Display for ValueKind {
-    fn fmt(&self, f: &mut std::fmt::Formatter<'_>) -> std::fmt::Result {
-        if f.alternate() {
-            f.write_str(self.as_natural())
-        } else {
-            f.write_str(self.as_str())
-        }
-    }
-}
-
-/// A massive function that takes in a type registration and the world and then
-/// does all the hard work of printing out the type nicely.
-fn fancy_debug_print(
-    resource: &IntoResource,
-    world: &World,
-    registrations: &[&TypeRegistration],
-) -> String {
-    const TAB: &str = "    ";
-    let registration = registrations.create_registration(resource.id);
-    let dyn_reflect = resource.ref_dyn_reflect(world, registration);
-
-    let reflect = dyn_reflect.reflect_path(resource.path.as_str()).unwrap();
-
-    fn debug_subprint(reflect: &dyn PartialReflect, indentation: usize) -> String {
-        let mut f = String::new();
-        let reflect_ref = reflect.reflect_ref();
-        let indentation_string = TAB.repeat(indentation);
-        match reflect_ref {
-            ReflectRef::Struct(struct_info) => {
-                f += "{\n";
-                for i in 0..struct_info.field_len() {
-                    let field = struct_info.field_at(i).unwrap();
-                    let field_name = struct_info.name_at(i).unwrap();
-
-                    let field_value = debug_subprint(field, indentation + 1);
-                    f += &format!(
-                        "{indentation_string}{TAB}{field_name}: {} = {field_value},\n",
-                        field.reflect_short_type_path(),
-                    );
-                }
-                f += &indentation_string;
-                f += "}";
-            }
-            ReflectRef::TupleStruct(tuple_struct) => {
-                f += "(";
-                f += &tuple_struct
-                    .iter_fields()
-                    .map(|field| debug_subprint(field, indentation + 1))
-                    .collect::<Vec<_>>()
-                    .join(", ");
-                f += ")";
-            }
-            ReflectRef::Tuple(tuple_info) => {
-                // f += "(\n";
-                // for field in tuple_info.iter_fields() {
-                //     let field_value = debug_subprint(field, indentation + 1);
-                //     f += &format!("{indentation_string}{TAB}{field_value},\n");
-                // }
-                // f += &indentation_string;
-                // f += ")";
-                f += "(";
-                f += &tuple_info
-                    .iter_fields()
-                    .map(|field| debug_subprint(field, indentation + 1))
-                    .collect::<Vec<_>>()
-                    .join(", ");
-                f += ")";
-            }
-            ReflectRef::List(_) => todo!(),
-            ReflectRef::Array(_) => todo!(),
-            ReflectRef::Map(_) => todo!(),
-            ReflectRef::Enum(variant) => {
-                // Print out the enum types
-                f += variant.variant_name();
-
-                match variant.variant_type() {
-                    VariantType::Struct => {
-                        f += " {\n";
-                        for field in variant.iter_fields() {
-                            f += &format!(
-                                "{indentation_string}{TAB}{}: {} = {},\n",
-                                field.name().unwrap(),
-                                field.value().reflect_short_type_path(),
-                                debug_subprint(field.value(), indentation + 1)
-                            );
-                        }
-                        f += &indentation_string;
-                        f += "}";
-                    }
-                    VariantType::Tuple => {
-                        f += "(\n";
-                        for field in variant.iter_fields() {
-                            f += &format!(
-                                "{indentation_string}{TAB}{} = {},\n",
-                                field.value().reflect_short_type_path(),
-                                debug_subprint(field.value(), indentation + 1)
-                            );
-                        }
-                        f += &indentation_string;
-                        f += ")";
-                    }
-                    VariantType::Unit => {}
-                }
-            }
-            ReflectRef::Opaque(reflect) => {
-                let Some(TypeInfo::Opaque(opaque)) = reflect.get_represented_type_info() else {
-                    return format!("{reflect:?}");
-                };
-                if opaque.is::<String>() {
-                    return format!("{GREEN}{reflect:?}{GREEN:#}");
-                } else {
-                    return format!("{YELLOW}{reflect:?}{YELLOW:#}");
-                }
-            }
-            ReflectRef::Set(_) => todo!(),
-        }
-
-        f
-    }
-
-    let mut f = String::new();
-    let reflect_ref = reflect.reflect_ref();
-    match reflect_ref {
-        ReflectRef::Struct(struct_info) => {
-            f += &format!("struct {} {{\n", struct_info.reflect_short_type_path());
-            for i in 0..struct_info.field_len() {
-                let field = struct_info.field_at(i).unwrap();
-                let field_name = struct_info.name_at(i).unwrap();
-
-                let field_value = debug_subprint(field, 1);
-                f += &format!(
-                    "{TAB}{RED}{}{RED:#}: {BRIGHT_YELLOW}{}{BRIGHT_YELLOW:#} = {},\n",
-                    field_name,
-                    field.reflect_short_type_path(),
-                    field_value
-                );
-            }
-            f += "}";
-        }
-        ReflectRef::Enum(set_variant_info) => {
-            // Print out the enum types
-            f += &format!("enum {} {{\n", set_variant_info.reflect_short_type_path());
-            let TypeInfo::Enum(enum_info) = registration.type_info() else {
-                unreachable!("{:?}", registration.type_info())
-            };
-            for variant in enum_info.iter() {
-                f += "\t";
-                f += variant.name();
-                match variant {
-                    VariantInfo::Struct(variant) => {
-                        f += " {\n";
-                        for field in variant.iter() {
-                            f += &format!(
-                                "{TAB}{TAB}{}: {},\n",
-                                field.name(),
-                                field.type_path_table().short_path()
-                            );
-                        }
-                        f += TAB;
-                        f += "}";
-                    }
-                    VariantInfo::Tuple(variant) => {
-                        f += "(";
-                        let mut iter = variant.iter();
-                        if let Some(first) = iter.next() {
-                            f += &format!("{}", first.type_path_table().short_path());
-                            for field in iter {
-                                f += &format!(", {}", field.type_path_table().short_path());
-                            }
-                        }
-                        f += ")";
-                    }
-                    VariantInfo::Unit(_) => {}
-                }
-                f += ",\n";
-            }
-            // Print out the current value
-            f += "} = ";
-            f += set_variant_info.variant_name();
-            match set_variant_info.variant_type() {
-                VariantType::Struct => {
-                    f += " {\n";
-                    for field in set_variant_info.iter_fields() {
-                        f += &format!("{TAB}{}: {:?},\n", field.name().unwrap(), field.value());
-                    }
-                    f += "}";
-                }
-                VariantType::Tuple => {
-                    f += "(\n";
-                    for field in set_variant_info.iter_fields() {
-                        f += &format!("{TAB}{:?},\n", field.value());
-                    }
-                    f += ")";
-                }
-                VariantType::Unit => {}
-            }
-        }
-        ReflectRef::Opaque(value) => {
-            f += &format!("{value:?}");
-        }
-        _ => f += &debug_subprint(reflect, 1),
-    }
-    f
-}
-
-impl From<()> for Value {
-    fn from((): ()) -> Self {
-        Value::None
-    }
 }
 
 macro_rules! from_t {
-    (impl $type:ty: $var:ident => $expr:expr) => {
+    (impl $type:ty: $var:pat => $expr:expr) => {
         impl From<$type> for Value {
             fn from($var: $type) -> Self {
                 $expr
@@ -520,8 +116,8 @@ macro_rules! from_number {
     };
 }
 
+from_t!(impl (): () => Value::None);
 from_number!(u8, u16, u32, u64, usize, i8, i16, i32, i64, isize, f32, f64);
-
 from_t!(impl String: string => Value::String(string));
 from_t!(impl bool: bool => Value::Boolean(bool));
 from_t!(impl Number: number => Value::Number(number));
@@ -532,195 +128,3 @@ from_t!(impl HashMap<String, Value>: hashmap => Value::Object(
         .map(|(k, v)| (k, UniqueRc::new(v)))
         .collect(),
 ));
-
-impl FunctionParam for Spanned<Value> {
-    type Item<'world, 'env, 'reg> = Self;
-    const IS_ARGUMENT: bool = true;
-
-    fn get<'world, 'env, 'reg>(
-        value: Option<Spanned<Value>>,
-        _: &mut Option<&'world mut World>,
-        _: &mut Option<&'env mut Environment>,
-        _: &'reg [&'reg TypeRegistration],
-    ) -> Result<Self::Item<'world, 'env, 'reg>, EvalError> {
-        Ok(value.unwrap())
-    }
-}
-impl<T: TryFrom<Spanned<Value>, Error = EvalError>> FunctionParam for Spanned<T> {
-    type Item<'world, 'env, 'reg> = Self;
-    const IS_ARGUMENT: bool = true;
-
-    fn get<'world, 'env, 'reg>(
-        value: Option<Spanned<Value>>,
-        _: &mut Option<&'world mut World>,
-        _: &mut Option<&'env mut Environment>,
-        _: &'reg [&'reg TypeRegistration],
-    ) -> Result<Self::Item<'world, 'env, 'reg>, EvalError> {
-        let value = value.unwrap();
-        Ok(Spanned {
-            span: value.span.clone(),
-            value: T::try_from(value)?,
-        })
-    }
-}
-impl FunctionParam for Value {
-    type Item<'world, 'env, 'reg> = Self;
-    const IS_ARGUMENT: bool = true;
-
-    fn get<'world, 'env, 'reg>(
-        value: Option<Spanned<Value>>,
-        _: &mut Option<&'world mut World>,
-        _: &mut Option<&'env mut Environment>,
-        _: &'reg [&'reg TypeRegistration],
-    ) -> Result<Self::Item<'world, 'env, 'reg>, EvalError> {
-        Ok(value.unwrap().value)
-    }
-}
-
-macro_rules! impl_function_param_for_value {
-    (impl $type:ty: $value_kind:ident($value_name:ident) => $return:expr) => {
-        impl_function_param_for_value!(impl $type: $value_kind($value_name) {$value_kind} => $return);
-    };
-    (impl $type:ty: $value_kind:ident($value_name:ident) {$kind:ident} => $return:expr) => {
-        impl FunctionParam for $type {
-            type Item<'world, 'env, 'reg> = Self;
-            const IS_ARGUMENT: bool = true;
-
-            fn get<'world, 'env, 'reg>(
-                value: Option<Spanned<Value>>,
-                _: &mut Option<&'world mut World>,
-                _: &mut Option<&'env mut Environment>,
-                _: &'reg [&'reg TypeRegistration],
-            ) -> Result<Self::Item<'world, 'env, 'reg>, EvalError> {
-                let Spanned { span, value } = value.unwrap();
-                if let Value::$value_kind($value_name) = value {
-                    Ok($return)
-                } else {
-                    Err(EvalError::IncorrectFunctionParameterType {
-                        expected: ValueKind::$kind,
-                        actual: value.kind(),
-                        span,
-                    })
-                }
-            }
-        }
-        impl TryFrom<Spanned<Value>> for $type {
-            type Error = EvalError;
-
-            fn try_from(Spanned { span, value }: Spanned<Value>) -> Result<Self, Self::Error> {
-                if let Value::$value_kind($value_name) = value {
-                    Ok($return)
-                } else {
-                    Err(EvalError::IncorrectFunctionParameterType {
-                        expected: ValueKind::$kind,
-                        actual: value.kind(),
-                        span,
-                    })
-                }
-            }
-        }
-    };
-}
-macro_rules! impl_function_param_for_numbers {
-    ($generic:ident ($($number:ident),*$(,)?)) => {
-        $(
-            impl FunctionParam for $number {
-                type Item<'world, 'env, 'reg> = Self;
-                const IS_ARGUMENT: bool = true;
-
-                fn get<'world, 'env, 'reg>(
-                    value: Option<Spanned<Value>>,
-                    _: &mut Option<&'world mut World>,
-                    _: &mut Option<&'env mut Environment>,
-                    _: &'reg [&'reg TypeRegistration],
-                ) -> Result<Self::Item<'world, 'env, 'reg>, EvalError> {
-                    let Spanned { span, value } = value.unwrap();
-                    match value {
-                        Value::Number(Number::$number(value)) => Ok(value),
-                        Value::Number(Number::$generic(value)) => Ok(value as $number),
-                        _ => Err(EvalError::IncorrectFunctionParameterType {
-                            expected: ValueKind::Number(NumberKind::$number),
-                            actual: value.kind(),
-                            span,
-                        })
-                    }
-                }
-            }
-            impl TryFrom<Spanned<Value>> for $number {
-                type Error = EvalError;
-
-                fn try_from(Spanned {span, value}: Spanned<Value>) -> Result<Self, Self::Error> {
-                    match value {
-                        Value::Number(Number::$number(value)) => Ok(value),
-                        Value::Number(Number::$generic(value)) => Ok(value as $number),
-                        _ => Err(EvalError::IncorrectFunctionParameterType {
-                            expected: ValueKind::Number(NumberKind::$number),
-                            actual: value.kind(),
-                            span
-                        })
-                    }
-                }
-            }
-        )*
-    };
-}
-
-impl_function_param_for_numbers!(Float(f32, f64));
-impl_function_param_for_numbers!(Integer(u8, u16, u32, u64, usize, i8, i16, i32, i64, isize));
-
-impl_function_param_for_value!(impl bool: Boolean(boolean) => boolean);
-impl_function_param_for_value!(impl Number: Number(number) {AnyNumber} => number);
-impl_function_param_for_value!(impl String: String(string) => string);
-// impl_function_param_for_value!(impl HashMap<String, UniqueRc<Value>>: Value::Object(object) => object);
-impl_function_param_for_value!(impl HashMap<String, Value>: Object(object) => {
-    object.into_iter().map(|(k, v)| (k, v.into_inner())).collect()
-});
-impl_function_param_for_value!(impl StrongRef<Value>: Reference(reference) => reference.upgrade().unwrap());
-
-impl FunctionParam for &mut World {
-    type Item<'world, 'env, 'reg> = &'world mut World;
-    const IS_ARGUMENT: bool = false;
-
-    fn get<'world, 'env, 'reg>(
-        _: Option<Spanned<Value>>,
-        world: &mut Option<&'world mut World>,
-        _: &mut Option<&'env mut Environment>,
-        _: &'reg [&'reg TypeRegistration],
-    ) -> Result<Self::Item<'world, 'env, 'reg>, EvalError> {
-        let Some(world) = world.take() else {
-            // make this unreachable by checking the function when it gets registered
-            todo!("world borrowed twice");
-        };
-
-        Ok(world)
-    }
-}
-
-// This probably isn't a good idea. But eh who cares, more power to the user.
-impl FunctionParam for &mut Environment {
-    type Item<'world, 'env, 'reg> = &'env mut Environment;
-    const IS_ARGUMENT: bool = false;
-
-    fn get<'world, 'env, 'reg>(
-        _: Option<Spanned<Value>>,
-        _: &mut Option<&'world mut World>,
-        environment: &mut Option<&'env mut Environment>,
-        _: &'reg [&'reg TypeRegistration],
-    ) -> Result<Self::Item<'world, 'env, 'reg>, EvalError> {
-        Ok(environment.take().unwrap())
-    }
-}
-
-impl FunctionParam for &[&TypeRegistration] {
-    type Item<'world, 'env, 'reg> = &'reg [&'reg TypeRegistration];
-    const IS_ARGUMENT: bool = false;
-
-    fn get<'world, 'env, 'reg>(
-        _: Option<Spanned<Value>>,
-        _: &mut Option<&'world mut World>,
-        _: &mut Option<&'env mut Environment>,
-        registrations: &'reg [&'reg TypeRegistration],
-    ) -> Result<Self::Item<'world, 'env, 'reg>, EvalError> {
-        Ok(registrations)
-    }
-}
